Move prove and verify to Stark module
